+++ app.js
@@ -25,13 +25,13 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
+app.use('/static', express.static('/data/static'));
 
 app.use('/weixin', proxy('http://59.57.240.50:8090/', {
   proxyReqPathResolver: req => {
     return `/weixin/${url.parse(req.url).path}`;
   },
 }));
-
 
 app.get('/', (req, res) => {
   res.render('../public/index');