+++ utils/catchMongooseError.js
@@ -3,8 +3,9 @@ const CaaError = require('./CaaError');
 module.exports = err => {
   if (!err) return null;
 
+
   let key = _.chain(err.errors).keys().first().value();
-  let error = err.errors[key];
+  let error = key ? err.errors[key] : null;
 
   return CaaError(500, error ? error.message : err.message);
 };